Add additional class on div if object is marked for destruction
In order to choose how to handle, after validation errors,
fields that are marked for destruction, add an additional class
on the div if the object is marked for destruction. This allows
me to be sure the div is hidden. Or, I could use the additional
class to style the div and its contents if I prefer it to be
visible.